+++ src/components/Contacts/ContactDetails/ContactTasksTab/ContactTaskRow/ContactTaskRow.stories.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from 'react';
-import { ResultEnum } from '../../../../../graphql/types.generated';
-import { gqlMock } from '../../../../../__tests__/util/graphqlMocking';
+import { ResultEnum } from '../../../../../../graphql/types.generated';
+import { gqlMock } from '../../../../../../__tests__/util/graphqlMocking';
 import { ContactTaskRow } from './ContactTaskRow';
 import {
   ContactTaskRowFragment,
