+++ src/components/Contacts/ContactDetails/ContactTasksTab/ContactTaskRow/TaskCommentsButton/TaskCommentsButton.test.tsx
@@ -2,7 +2,7 @@ import { MuiThemeProvider } from '@material-ui/core';
 import { render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
-import theme from '../../../../../theme';
+import theme from '../../../../../../theme';
 import { TaskCommentsButton } from './TaskCommentsButton';
 
 const numberOfComments = 10;
